backend:refactor: views->utils
