+++ backend/linklink/views_linklink/views_friend_request.py
@@ -17,7 +17,7 @@ from ..models import (
     LinkLinkUser,
     FriendRequest,
 )
-from ..views import is_within_twochon
+from ..utils import is_within_twochon
 
 #--------------------------------------------------------------------------
 #   FriendRequest Related APIs
